# GGRC-5634 Improve performance of fetching Assessments through QueryAPI