--- a/src/ggrc/models/mixins/issue_tracker.py
+++ b/src/ggrc/models/mixins/issue_tracker.py
@@ -54,6 +54,14 @@ class IssueTracked(object):
         uselist=False,
     )
 
+  @classmethod
+  def eager_query(cls):
+    """Define fields to be loaded eagerly to lower the count of DB queries."""
+    query = super(IssueTracked, cls).eager_query()
+    return query.options(
+        orm.subqueryload('issuetracker_issue')
+    )
+
   @simple_property
   def issue_tracker(self):
     """Returns representation of issue tracker related info as a dict."""