@Tobichimaru 2018-08-13
Adding `missing tests` because it would be nice to have a test for this in order not to check it by hands in the future