@Tobichimaru 2018-08-15
Warning: the code style will fail because of code duplication of list of factories in the integration test.