@Tobichimaru 2018-09-11
**Summary:** 
1. The code style is failing because of code duplication in the test. It is intended.
2. test_total_reindex is not in scope of this ticket, it just make it work again. It turned out that for the models CycleTaskEntry and CycleTaskGroupObjectTask it does not work properly since count of queries includes not only this models reindex but also its parent objects. I've added some TODO items in the test, but I think It should not be fixed in scope of this ticket.